refactor(biome): disable `noBarrelFile` for index files
