--- examples/nextjs-api-reference/app/client/components/index.ts
@@ -1,2 +1,0 @@
-export * from './Button'
-export * from './ClientWrapper'
