+++ integrations/nextjs/src/index.ts
@@ -1,2 +1,1 @@
-// biome-ignore lint/performance/noBarrelFile: package entrypoint
 export * from './ApiReference'
