+++ packages/api-client/src/v2/blocks/scalar-address-bar-block/index.ts
@@ -1,3 +1,2 @@
-// biome-ignore lint/performance/noBarrelFile: It's a block
 export { default as AddressBar } from './components/AddressBar.vue'
 export type { History } from './components/AddressBarHistory.vue'
