+++ packages/api-client/src/v2/blocks/scalar-auth-selector-block/index.ts
@@ -1,2 +1,1 @@
-// biome-ignore lint/performance/noBarrelFile: It is a block
 export { default as AuthSelector } from './components/AuthSelector.vue'
