+++ packages/api-reference/src/blocks/scalar-auth-selector-block/index.ts
@@ -1,2 +1,1 @@
-// biome-ignore lint/performance/noBarrelFile: <explanation>
 export { default as AuthSelector } from './components/AuthSelector.vue'
