+++ packages/api-reference/src/blocks/scalar-client-selector-block/index.ts
@@ -1,2 +1,1 @@
-// biome-ignore lint/performance/noBarrelFile: <explanation>
 export { default as ClientSelector } from './components/ClientSelector.vue'
