+++ packages/api-reference/src/blocks/scalar-info-block/index.ts
@@ -1,3 +1,2 @@
-// biome-ignore lint/performance/noBarrelFile: <explanation>
 export { default as InfoBlock } from './components/InfoBlock.vue'
 export { default as IntroductionCardItem } from './components/IntroductionCardItem.vue'
