+++ packages/api-reference/src/blocks/scalar-server-selector-block/index.ts
@@ -1,2 +1,1 @@
-// biome-ignore lint/performance/noBarrelFile: It's a block
 export { default as ServerSelector } from './components/ServerSelector.vue'
