+++ packages/api-reference/src/features/specification-extension/index.ts
@@ -1,3 +1,2 @@
-// biome-ignore lint/performance/noBarrelFile: <explanation>
-export { default as SpecificationExtension } from './SpecificationExtension.vue'
 export { getXKeysFromObject } from './helpers'
+export { default as SpecificationExtension } from './SpecificationExtension.vue'
