+++ packages/json-magic/src/dereference/index.ts
@@ -1,2 +1,1 @@
-// biome-ignore lint/performance/noBarrelFile: entrypoint
 export { dereference } from './dereference'
