+++ packages/json-magic/src/magic-proxy/index.ts
@@ -1,2 +1,1 @@
-// biome-ignore lint/performance/noBarrelFile: entrypoint
 export { createMagicProxy, getRaw } from './proxy'
