+++ packages/openapi-to-markdown/src/index.ts
@@ -1,2 +1,1 @@
-export { createMarkdownFromOpenApi } from './create-markdown-from-openapi'
-export { createHtmlFromOpenApi } from './create-markdown-from-openapi'
+export { createHtmlFromOpenApi, createMarkdownFromOpenApi } from './create-markdown-from-openapi'
