+++ packages/openapi-upgrader/src/index.ts
@@ -1,2 +1,1 @@
-// biome-ignore lint/performance/noBarrelFile: root entry point
 export { upgrade } from './upgrade'
