+++ packages/types/src/index.ts
@@ -1,7 +1,6 @@
 /**
  * We should not use these exports anymore, but we need them for commonjs compatibility.
  */
-/** biome-ignore-all lint/performance/noBarrelFile: entrypoint */
 
 // biome-ignore lint/performance/noReExportAll: leave this to avoid copy exports
 export * from './api-reference/index'
