+++ packages/workspace-store/src/events/index.ts
@@ -1,4 +1,3 @@
-// biome-ignore lint/performance/noBarrelFile: Entry point for /events
 export { type WorkspaceEventBus, createWorkspaceEventBus } from './bus'
 export type { ApiReferenceEvents, CollectionType } from './definitions'
 export { onCustomEvent } from './listeners'
