+++ packages/workspace-store/src/navigation/index.ts
@@ -1,4 +1,3 @@
-// biome-ignore lint/performance/noBarrelFile: Entry point for /navigation
 export { getOpenapiObject } from './helpers/get-openapi-object'
 export { getParentEntry } from './helpers/get-parent-entry'
 export { traverseDocument as createNavigation } from './helpers/traverse-document'
